add activesupport from rails, bz2 build script
git-svn-id: svn+ssh://burned.com/home/63904/data/svn/buildphp/trunk@7 34df37a4-c167-0410-bca2-a9e25f3560c2

+++ lib/build_task_abstract.rb
@@ -1,10 +1,12 @@
 class BuildTaskAbstract
   @prefix = '[...] '
   @config = {}
+  @install_dir = File.join(Dir.pwd, 'local')
   
   class << self
     attr_accessor :prefix
     attr_accessor :config
+    attr_accessor :install_dir
 
     def stop(msg)
       abort prefix + msg
@@ -18,52 +20,60 @@ class BuildTaskAbstract
       './configure'
     end
     
-    def get(task)
-      config[:package][:path] = File.join(ENV['BUILDPHP_EXTRACT_TO'], config[:package][:name])
+    def get()
+      if not is_installed
+        config[:package][:path] = File.join(ENV['BUILDPHP_EXTRACT_TO'], config[:package][:name])
 
-      Dir.chdir(ENV['BUILDPHP_EXTRACT_TO'])
+        Dir.chdir(ENV['BUILDPHP_EXTRACT_TO'])
 
-      if File.exists?(config[:package][:path]) && config[:package][:md5] == Digest::MD5.file(config[:package][:path]).hexdigest
-        notice "package already downloaded"
-      else
-        # get
-        sh "rm -f #{config[:package][:name]} && wget #{config[:package][:location]}"
-        # verify
-        stop 'MD5 does not match' if config[:package][:md5] != Digest::MD5.file(config[:package][:path]).hexdigest
+        if File.exists?(config[:package][:path]) && config[:package][:md5] == Digest::MD5.file(config[:package][:path]).hexdigest
+          notice "package already downloaded"
+        else
+          # get
+          sh "rm -f #{config[:package][:name]} && wget #{config[:package][:location]}"
+          # verify
+          stop 'MD5 does not match' if config[:package][:md5] != Digest::MD5.file(config[:package][:path]).hexdigest
+        end
+
+        # remove folder if already extracted.
+        sh "rm -rf #{config[:extract][:dir]}" if File.exists?(config[:extract][:dir])
+
+        # extract archive
+        sh "tar xfz #{config[:package][:path]}"
       end
-
-      # remove folder if already extracted.
-      sh "rm -rf #{config[:extract][:dir]}" if File.exists?(config[:extract][:dir])
-
-      # extract archive
-      sh "tar xfz #{config[:package][:path]}"
     end # /get
     
-    def configure(task)
-      stop "extract folder does not exist" if not File.exists?(config[:extract][:dir])
-      Dir.chdir(config[:extract][:dir]) do
-        sh get_build_string() do |ok,res|
-          stop "configure failed" if not ok
+    def configure()
+      if not is_installed
+        stop "extract folder does not exist" if not File.exists?(config[:extract][:dir])
+        Dir.chdir(config[:extract][:dir]) do
+          sh get_build_string() do |ok,res|
+            stop "configure failed" if not ok
+          end
         end
       end
     end # /configure
     
-    def compile(task)
-      Dir.chdir(config[:extract][:dir]) do
-        sh "make" do |ok,res|
-          stop "build failed" if not ok
+    def compile(cmd = 'make')
+      if not is_installed
+        Dir.chdir(config[:extract][:dir]) do
+          sh cmd do |ok,res|
+            stop "build failed" if not ok
+          end
         end
       end
     end # /compile
     
-    def install(task)
+    def install(cmd = 'make install')
       if not is_installed
         Dir.chdir(config[:extract][:dir]) do
-          sh "make install" do |ok,res|
+          sh cmd do |ok,res|
             stop "install failed" if not ok
             notice "install succeeded!"
           end
         end
+      else
+        notice "already installed"
       end
     end # /install
   end
